Revert "Revert "a first draft of PocCreationLoop""
