+++ src/main/resources/application-docker.conf
@@ -26,6 +26,7 @@ system {
     goClientURL = ${GO_CLIENT_URL_REGISTER_DEVICE}
     goClientToken = ${GO_CLIENT_TOKEN}
     thingApiURL = ${THING_API_URL_CREATE_DEVICE_PER_TENANT}
+    pocCreationInterval = ${POC_CREATION_INTERVAL}
   }
 
   aesEncryption {
