+++ src/main/resources/application.base.conf
@@ -26,6 +26,7 @@ system {
     goClientURL = "not reachable from outside"
     goClientToken = "goClientToken"
     thingApiURL = "https://api.console.dev.ubirch.com/ubirch-web-ui/api/v1/devices/create?with_api_info=true"
+    pocCreationInterval = 2
   }
 
   aesEncryption {
