+++ src/main/scala/com/ubirch/Binder.scala
@@ -134,6 +134,8 @@ class Binder extends AbstractModule {
   def InformationProvider: ScopedBindingBuilder =
     bind(classOf[InformationProvider]).to(classOf[InformationProviderImpl])
 
+  def PocCreator: ScopedBindingBuilder = bind(classOf[PocCreator]).to(classOf[PocCreatorImpl])
+
   def PocCreationLoop: ScopedBindingBuilder = bind(classOf[PocCreationLoop]).to(classOf[PocCreationLoopImpl])
 
   override def configure(): Unit = {
@@ -175,6 +177,7 @@ class Binder extends AbstractModule {
     KeyHashVerifier
     HashingService
     HashKeyRepository
+    PocCreator
     PocCreationLoop
     ()
   }
