+++ src/test/resources/application.conf
@@ -1,5 +1,11 @@
 include "application.base.conf"
 
+system {
+  services {
+    pocCreationInterval = 2
+  }
+}
+
 keycloak-users {
   polling {
     interval = 1
