+++ src/test/scala/com/ubirch/services/poc/PocTestHelper.scala
@@ -1,0 +1,60 @@
+package com.ubirch.services.poc
+import com.ubirch.Awaits
+import com.ubirch.ModelCreationHelper.{ createPoc, createPocStatus, createTenant }
+import com.ubirch.db.tables.{ PocStatusTestTable, PocTestTable, TenantTestTable }
+import com.ubirch.models.keycloak.group.{ CreateKeycloakGroup, GroupId, GroupName }
+import com.ubirch.models.poc.{ Poc, PocStatus }
+import com.ubirch.models.tenant.{ Tenant, TenantDeviceGroupId, TenantUserGroupId }
+import com.ubirch.services.keycloak.groups.TestKeycloakGroupsService
+import com.ubirch.services.{ DeviceKeycloak, UsersKeycloak }
+import com.ubirch.util.ServiceConstants.TENANT_GROUP_PREFIX
+import monix.execution.Scheduler.Implicits.global
+import org.scalatest.Assertion
+import org.scalatest.Matchers.convertToAnyShouldWrapper
+
+import scala.concurrent.duration.DurationInt
+
+object PocTestHelper extends Awaits {
+
+  def createPocTriple(): (Poc, PocStatus, Tenant) = {
+    val tenant = createTenant()
+    val poc = createPoc(tenantName = tenant.tenantName)
+    val pocStatus = createPocStatus(poc.id)
+    (poc, pocStatus, tenant)
+  }
+
+  def createNeededTenantGroups(tenant: Tenant, groups: TestKeycloakGroupsService): Tenant = {
+    val tenantGroup = CreateKeycloakGroup(GroupName(TENANT_GROUP_PREFIX + tenant.tenantName.value))
+    val deviceGroup: GroupId = await(groups.createGroup(tenantGroup, DeviceKeycloak), 1.seconds).right.get
+    val userGroup: GroupId = await(groups.createGroup(tenantGroup, UsersKeycloak), 1.seconds).right.get
+    tenant.copy(
+      deviceGroupId = TenantDeviceGroupId(deviceGroup.value),
+      userGroupId = TenantUserGroupId(userGroup.value))
+  }
+
+  def addPocTripleToRepository(
+    tenantTable: TenantTestTable,
+    pocTable: PocTestTable,
+    pocStatusTable: PocStatusTestTable,
+    poc: Poc,
+    pocStatus: PocStatus,
+    updatedTenant: Tenant): Unit = {
+
+    await(tenantTable.createTenant(updatedTenant), 1.seconds)
+    await(pocTable.createPoc(poc), 1.seconds)
+    await(pocStatusTable.createPocStatus(pocStatus), 1.seconds)
+  }
+
+  def assertStatusAllTrue(status: PocStatus): Assertion = {
+    status.deviceRealmRoleCreated shouldBe true
+    status.deviceRealmGroupCreated shouldBe true
+    status.deviceRealmGroupRoleAssigned shouldBe true
+
+    status.userRealmRoleCreated shouldBe true
+    status.userRealmGroupCreated shouldBe true
+    status.userRealmGroupRoleAssigned shouldBe true
+    status.deviceCreated shouldBe true
+    status.goClientProvided shouldBe true
+    status.certApiProvided shouldBe true
+  }
+}
